now sporting stable and unstable variants. i am starting to think that development of dsniff 2.4 has been delayed indefinitely so 2.4b1 might forever be available as +unstable.
known issue when attempting to force libpcap to install. someone care to
explain why libpcap does not install and even more so why it isn't even
recognized as a dependency? until this is resolved, the user needs to
install libpcap by hand - this is no worse than the portfile previously
available.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@1150 d073be05-634f-4543-b044-5fe20cf6d1d6