Band aid for perl version vs mod_perl port conflict
Apache mod_perl needs to know what version of Perl is installed so it can use the correct path to fix installed files. If the user installs perl, then perl is updated (and apache is updated in sync) and the user then attempts to install apache +mod_perl it will file as the path specified in the portfile won't exist.

This is very crap. I need a method to discover version of perl that is being used to build mod_perl so I can create correct path.

Currently forcing specific version of perl to be installed

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@5700 d073be05-634f-4543-b044-5fe20cf6d1d6