Updated to release 1.4.4.
Release Notice:
Zope Corporation is pleased to announce the release of version 1.4.4 of the
Zope Content Management Framework (CMF). This release consists primarily of bug
fixes for version 1.4, released in April, 2003, and closes a potential e-mail
header injection vulnerability. All users of earlier CMF versions which surface
through-the-web registration to the public web should upgrade to version 1.4.4.

Submitted by:   rshaw

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@7005 d073be05-634f-4543-b044-5fe20cf6d1d6