New port for rdiff-backup Python script/package.
rdiff-backup is a script, written in python(1) that backs up one directory to another.  The target directory ends up a copy (mirror) of the source directory, but extra reverse diffs are stored in a special subdirectory of that target directory, so you can still recover files lost some time ago. The idea is to combine the best features of a mirror and an incremental backup. rdiff-backup also preserves symlinks, special files, hardlinks, permissions, uid/gid ownership, and modification times.

Submitted by:   rshaw

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@7277 d073be05-634f-4543-b044-5fe20cf6d1d6