Update php5 --> 5.0.5
According to the php news note:

[06-Sep-2005] PHP 5.0.5 is now available for download. This version is a
maintenance release, that contains numerous bug fixes, including security
fixes to vulnerabilities found in the XMLRPC package. All users of PHP 5.0
are encouraged to upgrade to this version.

The full list of changes in PHP 5.0.5 is available in the PHP 5 ChangeLog
at http://www.php.net/ChangeLog-5.php#5.0.5.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@14309 d073be05-634f-4543-b044-5fe20cf6d1d6