Submitted by:   jmpp@
Initial checkin of an updated DarwinPorts Portfile that now uses our numbered
releases instead of a CVS checkout.

This is meant primarily for people working on the dmg binary release to hack
on at will, so that in the end the entire process can be summarized with a
simple "port dmg DarwinPorts". This will either imply populating the files/
dir with the necessary files (localizations, pre/post-flight scripts, etc) or
hacking on our sources providing the "pkg" target to echieve the creation of
this enhanced type of packages, or both. In any case, the use of pre/post-{pkg,dmg}
hooks might be necessary.

The Portfile is just bare bones at the moment, variants can be added later on
to achieve user confined installations through our configure arguments, among
others.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@14628 d073be05-634f-4543-b044-5fe20cf6d1d6