Port:		simage Version:	1.6.1 Revision:	1 Bug:		https://trac.macports.org/ticket/7573
Update port graphics/simage
From the bug:  "SIM's simage library can configure using QuickTime for image
loading. This avoids the need to build a number of extra libraries, and it
can reduce the complexity of dependencies. The updated Portfile includes a
quicktime variant that removes the image-loading libraries that are not
needed when QuickTime handles the corresponding formats instead. I also
modified the dependencies to not explicitly require ports. The dependencies
do not explicitly require anything unique to DarwinPorts-installed versions
of those libraries.
This update also corrects a problem with the configure script looking for
incorrect paths when trying to locate libsndfile. The --with-libsndfile
configure option does not obey porthier. The attached patch modifies configure
to look in ${prefix}/include and ${prefix}/lib."

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@17422 d073be05-634f-4543-b044-5fe20cf6d1d6