Bug:            NONE Submitted by:   rhwood@ Reviewed by: Approved by: Obtained from:
New port: gnome-netstatus, a GNOME panel applet that shows network interface statuses.
NOTE: this port is currently useless, as it thinks that all network interface information is exported to the /proc filesystem/directory structure--which is true for MOST Unix/Linux/BSD systems. If someone wants to help the gnome folks get this working right, then please feel free to do so.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@18522 d073be05-634f-4543-b044-5fe20cf6d1d6