Use new configure.(ldflags|cflags)-append syntax. No longer override LDFLAGS, CFLAGS and CPPFLAGS. This also seems to fix the +universal variant.
And no longer use -O3. It's been there since the first version of
the port but I don't know why glib2 should be singled out for that
"optimization," if it even is one; all generated code files are
exactly the same size with either -O2 or -O3 so I'm getting rid of it.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@25012 d073be05-634f-4543-b044-5fe20cf6d1d6