devel/premake: Update silently changed distfile
For some reason, the premake distfile on sourceforge has been silently replaced
with another one.  The filename suggests a teeny version bump, but the website
and the files contained within the tarball do not.  For this reason, this patch
is being applied as a revision bump rather than a version bump.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@27282 d073be05-634f-4543-b044-5fe20cf6d1d6