offlineimap: update to latest version (5.99.4).
The changelog states that all versions released so far are unstable,
one user reports the latest version to be significantly faster, and
there are no open bugs against 5.99.4 that are closed against 4.0.16,
hence the decision to update this port despite the apparent release
candidate numbering.

In addition, lines have been removed from the Portfile that pertain to
files no longer in the tarball.

This closes https://trac.macports.org/ticket/13035.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@30705 d073be05-634f-4543-b044-5fe20cf6d1d6