cairo: update to 1.6.4
cairo 1.5.2 and later require an external pixman library; it's no longer included in the cairo source as it was before, so a dependency on libpixman has been added.

quartz support is no longer considered experimental in cairo 1.6 and is enabled by default, so the quartz and atsui variants have been removed.

pdf support is enabled by default so the pdf variant has been removed.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@36051 d073be05-634f-4543-b044-5fe20cf6d1d6