Quick reversion to address breakage if 'emacs' or 'emacs-app' is not installed:
From Ryan (about the top-level set forms):

Ah, but you can't do it this way. This relies on emacs or emacs-app
being installed. But it might not yet be installed at the time the
user runs "port info" or "port variants" or "port fetch" or "port
checksum" or "port deps" or "portindex" which will cause an error,
like it's doing for the PortIndexRegen.sh script now:

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@39992 d073be05-634f-4543-b044-5fe20cf6d1d6