pure update to 0.35 pure-devel: update to 0.35-r2302
All tests passed on Leopard/Intel.

Remove --enable-release because the INSTALL file encourages us not to use it "unless performance is really critical".

On PowerPC, use --disable-fastcc and require llvm 2.6 because the INSTALL file says to, noting that llvm 2.6 is not in MacPorts yet (see https://trac.macports.org/ticket/21174) but that the resulting inability to install Pure at all on PowerPC is not a problem since it has never worked properly on PowerPC before (most tests in the test suite failed).

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@57845 d073be05-634f-4543-b044-5fe20cf6d1d6