Create new port php5-gdchart
The GDChart extension provides an interface to the bundled gdchart
library. This library uses GD (bundled) library to generate 20 different
types of graphs based on supplied parameters. The extension provides an
OO interface to gdchart exposing majority of options via properties and
complex (array) options via a series of methods.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@62720 d073be05-634f-4543-b044-5fe20cf6d1d6