libnotify: Don't install on Lion
Yes, I know it installs and runs fine, but when libnotify.dylib is present in
${prefix}/lib, it causes problems when linking future binaries.  This results
in miscelaneous build failures and runtime failures that are difficult to
diagnose.

Rather than subject users to these problems, it is better to disable this port
until we can come up with a good solution for Lion.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@79664 d073be05-634f-4543-b044-5fe20cf6d1d6