clang-* llvm-*: New ports for multiple versions of llvm and clang
Like gcc, multiple versions of llvm and clang can now be installed
simultaneously.  These ports will go through a period of further testing and
bugfixing after which the existing ports (clang, clang-devel, llvm, and
llvm-devel) will be obsoleted.

These ports install binaries and man pages with the -mp-<version> suffix to
maintain consistency with the gcc ports.

A port select group will be added to allow choosing the version of choice, but
the naming and mechanics of that group are still undetermined and need to be
discussed on the mailing list.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@86535 d073be05-634f-4543-b044-5fe20cf6d1d6