copy the current state of the ghc port to hs-platform-ghc
A follow-up commit will change haskell-platform to depend on this port.
A different follow-up commit will update the ghc port to a newer version.
Doing this preserves the current state of the haskell platform in MacPorts
(whatever that is).

Haskell platform will probably be updated soon, too (also with
a hs-platform-ghc port)

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@96876 d073be05-634f-4543-b044-5fe20cf6d1d6