textmate2: revert to using system ruby for the build process
Previously the build required rb-json, which forced us to use MacPorts
ruby. Since upstream has since removed this requirement and the build
works fine with the Apple's ruby on the platforms supported by this
port, we don't need a build-time dependency on ruby.

This closes https://trac.macports.org/ticket/37602, although not in the way suggested.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@101463 d073be05-634f-4543-b044-5fe20cf6d1d6