ufc: update to 1.3.0
This also enforces installation with python27 instead of providing
variants. The reason for this is that I didn't realize the fenics project had
such interdependencies with C/C++ and python and they really target only one
python at a time. So, instead of making this a burden for me to maintain, I'll
just force one python version at a time.

We can come back to the decision later, if need be.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@115861 d073be05-634f-4543-b044-5fe20cf6d1d6