py-pyphant: replace PIL dependency by Pillow in imageprocessing subport (see https://trac.macports.org/ticket/44472)
The py??-pil dependency in py??-pyphant-imageprocessing
is replaced by py??-Pillow.
The changed dependency is also patched in the source in the setup.py script
and in the requires.txt file in the egg-info folder.

A new fallback variant (pil) is automatically added as default,
when py??-pil is still active.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@122800 d073be05-634f-4543-b044-5fe20cf6d1d6