rust: Use a Git checkout to fetch source.
Rust is a fast-moving project that currently expects its users to be on
the bleeding edge, so we should start providing frequent development
snapshots instead of releases, which go out of date very quickly.

We'll continue treating the snapshot compiler as a distfile so that it's
checksummed and mirrored as usual.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@123559 d073be05-634f-4543-b044-5fe20cf6d1d6