py-scipy: add __ACCELERATE__ to cpp; fixes https://trac.macports.org/ticket/46395
This change adds a CPP flag for building against the default blas/lapack
libraries provided by Apple. We also use the revbump to remove the dependency
on SuiteSparse which was removed upstream for this latest release.

In addition, there was a problem with not all files having read access which
was fixed as well.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@131020 d073be05-634f-4543-b044-5fe20cf6d1d6