mpi-1.0: fix silently failing to enforce mpi variants
Previously, the mpi portgroup would try to append a mpi variant to the
default_variants variable. Unfortunately, this would fail and silently be
ignored leading to build-time issues.

This patch changes that behavior to raise an error explaining that the user
must match the mpi variant. For example, if hdf5 is installed with +mpich, then
netcdf must be installed with +mpich.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@131494 d073be05-634f-4543-b044-5fe20cf6d1d6