muniversal-1.0: Set options with `option` proc (https://trac.macports.org/ticket/47388)
A portfile command sets its associated option to a list composed of its
arguments. Accessing the associated global variable then yields that
list. Thus, for a portfile option `foo`,

    foo 1 2 3
    set bar $foo
    foo $bar

does not restore `foo` to a list containing 1, 2, and 3. Instead, `foo`
is set to a list containing a list containing 1, 2, and 3. One way to
correct this is with `eval`:

    foo 1 2 3
    set bar $foo
    eval foo $bar

The use of `eval` is generally discouraged because it's very easy to
create subtle injection bugs/vulnerabilities.[1] Tcl 8.5 argument
expansion permits an alternative:

    foo 1 2 3
    set bar $foo
    foo {*}$bar

It's not particularly obvious when the double substitution[2] provided
by `eval`/`{*}` is required. This has caused much confusion in both base
and portfile code.

Another option (more palatable in base code) is to dispense with the
portfile command magic altogether and use the `option` proc directly.

    option foo {1 2 3}
    set bar [option foo]
    option foo $bar

The `option` proc behaves more or less like `set` and is thus easier to
understand, if more verbose.

This is the approach I've taken to handling the bugs reported in https://trac.macports.org/ticket/47388
and https://trac.macports.org/ticket/47420. The muniversal-1.0 portgroup looks nothing like a portfile
at this point, so I think the use of portfile commands just muddles the
reader's mental model of what's going on. The simplicity of `option` is
preferable here.

[1] http://wiki.tcl.tk/1017
[2] http://wiki.tcl.tk/1535

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@135550 d073be05-634f-4543-b044-5fe20cf6d1d6