lang/ccl: update to ccl-1.11
The documentation is no longer installed as it is no longer apparently
pre-formatted in a HTML file, instead needing a dance with Quicklisp
to generate.  I chose to get the port updated over figuring out a
recipe for documentation.  Parties interested in generated
documentation for CCL can view <http://trac.clozure.com/ccldoc>, and
figure out how to patch it into the current process.

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@143236 d073be05-634f-4543-b044-5fe20cf6d1d6