neomutt: update +db4, add +lmdb
Add a variant for the LMDB headercache backend and make it the default if no
backend was selected, since it's faster than the previous default gdbm.

Improve the variant handling for the headercache backends and remove the
incorrect conflict between +headercache and +db4.

Update the +db4 variant to use db48 instead of db44.

Remove unnecessary -L/usr/lib linker flag from +headercache.

Closes: https://trac.macports.org/ticket/52293

git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@152965 d073be05-634f-4543-b044-5fe20cf6d1d6