New port munge @0.5.12
munge is a tool to generate authentication credentials that can be passed from
system to system. Credentials are generated using a shared key, and include
generation time, UID, and GID.

munge is most noticeably used by the SLURM job-scheduler system.  Since SLURM
relies on munge, this needs to go in first.

munge can use either OpenSSL or libgcrypt for the actual cryptographic work.
The current Portfile defaults to OpenSSL (with libgcrypt available as a
variant), but that can be changed.

https://trac.macports.org/ticket/50421