New port for wallet @1.3
wallet is a Kerberos-authenticated file storage & retrieval engine, which has
support for automatically generating things like Kerberos keytabs and random
passwords. Authentication is via Kerberos, and client-server communication is
via remctl. The client is written in C, the server is Perl, and there is also
a Perl script that can be placed on Kerberos 5 KDCs to aid in keytab
retrieval.

I am attaching the proposed Portfile and twelve patch files. The patches are
needed because wallet's build & install process does not allow setting custom
Perl paths. All of the patches have been submitted upstream.

Closes: https://trac.macports.org/ticket/50421