openttd: 1.6.1, actually apply the patches I wrote
It seems when updating to 1.5.2 I wrote the patches required to fix the
build on modern OS X, even submitted them upstream, but then failed to
actually enable them in the Portfile.

Update one of the patches with a newer suggestion from the upstream
ticket, update to 1.6.1 and enable the patches.

Closes: https://trac.macports.org/ticket/52841