cmake-1.1 portgroup: generator support.
This revision adds support for one more of CMake's generators, and
allows ports to use the more efficient ninja tool. This is supposed to
make a significant difference for large projects like llvm and clang,
even for once-off builds.
As an additional improvement the default destroot target for the Makefile
generator is set to install/fast. This skips the implicit "make" step
that is otherwise always part of "make install", and this can reduce
the destroot time significantly.

Committed from host : Portia.local