afsctool: support compression on macOS 10.12
The undocumented f_type returned by statfs(2) for HFS+ filesystems has
changed to 23 in macOS 10.12 Sierra. Without a source release of XNU on
opensource.apple.com, I have to guess this is the intended change.

Unconditionally applying the patch as the change will still work with
older releases where f_type is 17 and it is unlikely 23 would refer to
anything else.

Closes: https://trac.macports.org/ticket/52463