git-cola: switch to pyqt4 by default, add +qt5
Running git-cola with Qt5 exposes some artifacts, which was at least
verified for 10.12 Sierra and 10.11 El Capitan. Switch to Qt4 as default
instead. The Qt5 version can still be enabled with the +qt5 variant.

In the same patch selecting a fixed pyqt version, remove a custom
addition to sys.path specific to Homebrew which is likely to mess up
running git-cola when it is installed in /usr/local.

Closes: https://trac.macports.org/ticket/52868