port:qca* : proposed adoption and introduction of Qt5 subports
The QCA ports have been without maintainer for a bit now, and lack the
possibility to install for Qt4 and Qt5 concurrently. That possibility is
essential for the future KF5 ports, so I propose that Marko and I adopt
this port and introduce the required changes. The ports as defined by
the current Portfile have seen thorough testing by both of us as well as
by a few "beta testers" of my KF5 ports.
I don't think there is any point in maintaining the Qt4 and Qt5 variants.
The current version also introduces a few changes that are unrelated to
Qt5/KF5 but aim to simplify dependencies by reducing the number of
subports. This is open for discussion w.r.t. the Qt4 versions. I don't
think there is much justification (for use with MacPorts) to build QCA
as a framework, but this could be reintroduced as a variant.

Closes: #59
Closes: https://trac.macports.org/ticket/50500