sfml: fix install name of dylibs + pkg-config sup.
Remove cmake code that sets the install name of the dylibs to @rpath, so
that linking to them works.

Will send a patch upstream to allow disabling this with an option.

Also install pkg-config files (this requires a minor patch, and I will
try to address this upstream as well.)