inkscape: remove outdated +gtk3 variant
The inkscape 0.92.x series is scheduled to be the last gtk2
inplementation.  Development in trunk has removed all gtk2 support
and is now concentrated on a full featured gtk3 ui implementation for the
0.93.x series.  Since it's unlikely that any of this gtk3 code will
be backported to the 0.92.x branch, the +gtk3 variant is no longer
of much use or interest.