alps: update to 2.3.0, use Boost port and review available variants.
Clang is better supported in the new release and the current Boost port is supported
by alps 2.3.0.
For matching the Boost compiler variants I needed to enable the compilers PortGroup
which conflicts with the mpi PortGroup, hence the manual definition of openmpi and
mpich variants.
Python 3 would also work, but it requires Boost >= 1.63.0.