_resources/port1.0/group/muniversal-1.0.tcl: fix some builds in trace mode for ports not overwriting depends_build.
In trace mode, port itself can happily check for and execute random
files, but the actual build process spawned later will be unable to work
with ${prefix}/bin/nm if it's not a proper dependency.

Given that we only set NM to ${prefix}/bin/nm if it already exists on
the file system, it must be installed at that time and the build
dependency doesn't add additional cruft.