muniversal PG: do not explicitly set NM
This code was originally meant as a workaround for a libtool error
when cross-compiling.
See https://github.com/macports/macports-ports/commit/92dd44074ab7bd2ff2fcf5805b26931b78110907

Based on experiments with a 10.5 VM, each port for which this was a
problem started using a newer version of libtool sometime in the last
few years, and the workaround is no longer necessary.