MacPorts: Keep sip-workaround dir in installers
MacPorts by default deletes empty directories from destroot, unless they
are in destroot.keepdirs (which creates a .turd_MacPorts file in the
directory).

Because MacPorts 2.4.0 introduced a $prefix/var/macports/sip-workaround
directory but did not add it to the MacPorts port, the installers that
were built for 2.4.0 with this port lack the sip-workaround directory.

The code for trace mode will create the directory as needed, but it will
not have the correct permissions of drwxrwxrwt, which leads to problems
in trace mode, such as

| sip_copy_proc: mkdir(/opt/local/var/macports/sip-workaround/503): Permission denied

Closes: https://trac.macports.org/ticket/53435