nodejs4: Fix build with libc++
nodejs4 originally assumed all versions of macOS use tr1. This is not
true with libc++, which is the default on Mavericks and later.

The original patch was conditionally applied on Mavericks and later. It
removed the use of tr1 in one place, but not in another place, so the
build failed on Mavericks and later, and presumably on earlier systems
configured to use libc++.

nodejs6 contains preprocessor directives that avoids using tr1 on
Mavericks and later. This would presumably still fail on earlier systems
configured to use libc++.

This new patch avoids tr1 when libc++ is in use, regardless of macOS
version, and is unconditionally applied.

Closes: https://trac.macports.org/ticket/53401