ntpsec: new port
From the homepage <https://www.ntpsec.org>:
> NTPsec project - a secure, hardened, and improved implementation of Network
> Time Protocol derived from NTP Classic, Dave Mills’s original.
>
> NTPsec, as its name implies, is a more secure NTP. Our goal is to deliver
> code that can be used with confidence in deployments with the most stringent
> security, availability, and assurance requirements.

This port is a drop-in replacement for ntp, with hardened internals. It should
behave identically in most circumstances, except that obsolete and/or insecure
features have been removed.

The current version of this port incorporates two newer patches from upstream
to fix a compilation error when leap second smearing is enabled.

The example file included in this port is the same example file included in
the ntp port, but with small adaptations for changes in ntpsec.