gcc*: Fix building against an SDK
With this change, gcc now builds when configure.sdkroot is an SDK.  The
intention is that the installed gcc should be identical to a build in which
configure.sdkroot is /, but GCC's build system is flawed and does not
properly implement their --with-build-sdkroot= configure option.

As a workaround, we are using --with-sdkroot= which causes the installed gcc
to look in the SDKROOT by default instead of / by default.  Hopefully GCC
will provide a fix for this bug, or someone who is able to look at GPLv3 code
can provide a fix for them.  In the mean time, this is better than failing
to build.

See https://trac.macports.org/ticket/53726
See https://gcc.gnu.org/bugzilla/show_bug.cgi?id=79885

Signed-off-by: Jeremy Huddleston Sequoia <jeremyhu@macports.org>