ghostscript: Fix libgs.dylib install_name
The ghostscript Makefiles changed with the recent update to no longer
use LDFLAGS_SO, but now use three variables {GS,PCL,XPS}_LDFLAGS_SO, so
the patch we previously applied to get the correct library ID load
command broke.

Adjust it to set GS_LDFLAGS_SO, which is the correct value to set.

Note that this does not revbump dependents that need a rev-bump. That
should be done in a follow-up commit.

Closes: #54019