volk*: reinstate patch for devel.
Add back in to devel the temporary patch to fix building on older compilers that don't handle dynamic type casting robustly; use correct type API for 'pow' function. The patch is upstream in 'maint' but we use 'master' here for the devel port. Rev-bump devel to take effect; already in release, so no need to rev-bump it.