xemacs: Change negative variant.
Previously the default was to install SUMO and mule, and the +no_sumo variant installed only the minimal efs and xemacs-base packages. In accordance with our policy to avoid negative variants, there is now a +sumo variant (which is the default) and without this, just efs and xemacs-base are installed. A legacy +no_sumo variant has been left for compatibility in upgrades.