recently upgraded gnome C++ bindings: use PortGroup cxx11
All of these ports now require at least C++11 and in some cases
C++14 and even C++17. Legacy versions of these ports may be necessary
for ports that use gtkmm (the gtk2 bindings) but upstream regards
the gtkmm as obsolete and expects all developers to use the latest
gtkmm3 bindings.