py-gnupg: New port
This adds a port for the python-gnupg version by Vinay Sajip, which is
called "python-gnupg" on PyPI and also packaged in Debian. It appears to
be the more popular version in general.
Alternatively, there is a fork by Isis Lovecruft called "gnupg" on PyPI.
To avoid name conflicts, this port could be called "py-python-gnupg",
but that would still mean a clash for Python module "gnupg".