caffe: update to 1de4ceb (20170817)
Port for caffe was old (2015), so this commit upgrades its source to
current master stable. It updates required patches so library will
install successfuly.
Dynamic libraries are amended, so they will point to the correct
libcaffe. Current Port, has an issue with python and caffe, that
requires manual recolocation of affected library.
Also, tests are added for both library and python module.