gpgme: Add +gnupg20 and +gnupg21 variants
GnuPG 2.1.x (port gnupg21) now installs as ${prefix}/bin/gpg instead of
${prefix}/bin/gpg2, which makes it conflict with this port due to the
dependency on path:bin/gnupg2 -- this forces the installation of gnupg2
but that is not satisfiable (gnupg2 conflicts with gnupg21).

This commit addresses this issue by adding two variants, gnupg20 and
gnupg21, which allow the specific version that this port will depend on.
If none is explicitly chosen, gnupg20 is selected to preserve the old
behaviour.

Closes: https://trac.macports.org/ticket/54749