portgroup cxx11 1.1: remove redundant black/whitelisting
It's pointless having both a blacklist and a whitelist, since if a whitelist
exists then only the compilers in it will be used. It's also pointless having
multiple compilers provided by ports in the whitelist, as the first available
listed compiler will be used, and those provided by a port are always
available.