ffmpeg-devel: update to git master as of 20170912
ABI compatible with current ffmpeg.

As of this snapshot, use of --enable-vda on 10.11+ causes a
configuration error citing missing vda dependencies. Allowing
ffmpeg to auto-configure in this situation allows the build
to proceed but the effect on a previously cited bug is unverified.

Relevent test results and/or suggestions are welcome. For further
details, see Portfile comments and https://github.com/mpv-player/mpv/issues/2299.