gdk-pixbuf2: Fix configure broken by upstream
Upstream inadvertently removed os_win32=no for the darwin platform,
causing various configure checks to be skipped, for example for libjpeg.
This resulted in missing pixbuf loaders which were not included in the
build.

Reverting the upstream change completely, as it is not clear to me why
we would not want to use shared-mime-info as before.

Upstream: https://bugzilla.gnome.org/show_bug.cgi?id=786167