libmd: New maintained upstream, version 0.0.0
The previous libmd upstream was old and seems unmaintained. As it was
missing support for modern algorithms such as SHA256, SHA384, and
SHA512, switch to a maintained upstream that is also packaged in many
Linux distributions.

There are no direct dependents of this port in our ports tree, so this
change should not cause any problems. Being based on the BSD libmd
implementations, the provided API functions should be compatible.