spice-gtk: upgrade to 0.34
I've added a +quartz variant and also a warning that the +x11 variant
is untested.

The +quartz variant forces the build system to build all C sources as
Objective-C due to some #include oddities. Yes, this is rather large
hammer for this particular nail, but Objective-C is supposed to be a
strict superset of C, so... it works...

See: https://trac.macports.org/ticket/54439