plumed:small fix
Just after release v2.3.3, we realized that a tiny fix was missing.
The bug is not critical, but the fix allows plumed to be combined
with latest gromacs2016.4 instead of older gromacs2016.3.

I here use the git hash just containing the bug fix in order to have the
correct version installed on macports.

Clearly, at next plumed release (2.3.4 or 2.4.0) this hack will disappear.