gegl: fix build with glib2 1.54+
As part of the old introspection patch, this port
patches gegl-enums.h.  This in turn causes the build
to try and regenerate gegl-enums.c using the glib-mkenums
utility provided as part of glib2's GObject API.

Unfortunately, for reasons that are still not clear, the
new python implementation of glib-mkenums provided in glib2
1.54+ chokes on this conversion.

As a work-around, this commit patches gegl-enums.c to match
the patched gegl-enums.h, as generated in earlier versions,
allowing the build to complete as before and without tempting
the build to try and use glib-mkenums.

Note that this problem is specific to gegl and not the newer
versions gegl-0.3 and gegl-devel which now include a properly
generated gegl-enums.c in the upstream git repo.

Fixes https://trac.macports.org/ticket/54932