inkscape: fix build in High Sierra
The current version of Inkscape fails with a build error due to an
ambigious call to abs(). Indeed, i->pathNext->id.objID is of type
unsigned int, for which abs() does not make much sense. Adding a cast
to 'long' fixes it.

The root cause of the bug is likely that the code isn't strictly valid
C++, and so the version of Clang included with Xcode 9 rejects it,
where earlier versions of Clang accepted it. Since this is a build
fix, we don't bump the revision.

(Based on suggestion by Michael Lass)

Fixes: https://trac.macports.org/ticket/54886