neomutt: update to 20171013 (autosetup, new variants)
This upstream release migrated to autosetup for configuration insted of
autotools (which is still available but is deprecated). The switches in
the Portfile were changed to conform with the new syntax.

The variant +debug was removed since the switch disappeared from
upstream ('--debug' still exists but it controls the level of autosetup
logging, not if debug symbols are included in the binaries).

The variant +ssl was removed and SSL is now mandatory for this port.

New variants +gss and +lua were introduced that expose the respective
build configuration switches.

A bug related to building the documentation (due to missing dependencies)
was fixed. Building the documentation is now optional and controlled by
the +doc variant in order to minimize the number of dependencies
required by default.

Closes: https://trac.macports.org/ticket/54219