neomutt: add +mutt variant
Upstream now installs as 'neomutt' to allow side-by-side installation
with 'mutt'. Since MacPorts currently does not provide a 'mutt' port (it
has been removed due to the lack of maintainer) and also to preserve the
previous behaviour, this commit introduces a +mutt variant, enabled by
default, that installs a symlink so neomutt can still be invoked as
'mutt'.

In the event that a 'mutt' port is reintroduced in the future, this
variant can be turned off to allow side-by-side installation of both
ports. (Most likely it will become the new default in that case.)