Fix build failure with MacVim 8.0.snapshot138
`mvim` no longer contains the absolute path to the MacVim Applications
directory. Instead, it discovers that location from the absolute
location of the `mvim` script -- the script in `/opt/local/bin/mvim` is
now a symlink. Thus, the patch step in the current Portfile fails.

In addition, the install prefix for MacVim should be set to
`/Applications/MacPorts/MacVim` using the `configure` script, so that
similar issues don't rear their head in the future.