poedit: Remove outdated port
poedit upstream is now at 2.0.4, years ahead of 1.5.7 provided by this
port. Attempts to update the port to a more current version, feature
complete with the official upstream version, were initially blocked by
missing dependencies in MacPorts. I also have not used the software in
a long time myself, so that adds lack of interest on my side.

Providing such an old and outdated version via MacPorts has no value and
I am therefore removing this port from the ports tree. If you are
interested in an update, feel free to revive the port and take the
maintainer role.

Closes: https://trac.macports.org/ticket/43772