neomutt: Disable duplicate ccache on command line
autosetup detects ccache and then adds it in front of the compiler
automatically, even if MacPorts already did the same. Disable the
auto-detection to use the compiler as passed by MacPorts.

Futhermore, if the ccache port was installed but not enabled in
macports.conf, the build would fail as write operations to
${prefix}/var/macports/build/.ccache/ were blocked by the sandbox.

See: https://github.com/macports/macports-ports/pull/938
Closes: https://github.com/macports/macports-ports/pull/941