qt5[5-7]: keep configure.cxxflags empty
Attempting to mix and match MacPorts settings and Qt settings
has lead to difficulty in the past.

This change is only needed on older versions of Qt because
they are the only ones that build on older OS versions.

No revbump since port either builds successfully or not at all.