moreutils: Re-enable building on Leopard and Tiger
Building on Tiger was disabled because parallel failed to build.
Now that we no longer include parallel in moreutils, this is no
longer a problem.

See: https://trac.macports.org/ticket/24279

No reason was given why building on Leopard was disabled in
1f118454893940ec44023e64246b3f9af5ade936. It builds fine for me
when the Snow Leopard patch is added.

See: https://trac.macports.org/ticket/53359