terminal-notifier: Update to version 2.0.0
This release comes as an app bundle, but as it is not meant to be used
directly, it will be hidden under libexec in the prefix hierarchy.
A wrapper script is necessary to launch terminal-notifier as it expects
the first argument in argv to point into the app bundle path.