ocaml-ocamlbuild: add new port (at version 0.12.0)
* Add new port
* Set me as the maintainer

ocamlbuild was unbundled from ocaml in version 4.03. Thus, ports
that require ocamlbuild will need this added to their build dependencies.

Note that one such port is ocaml-camlp4, which needs to be updated in
sync with ocaml itself. Thus, this must be added in order to complete
the update of ocaml to 4.05.0.