cmake: Force libc++ on Snow Leopard and newer
This does not cause problems on libstdc++ systems because cmake doesn't
depend on any C++ libraries, nor does it supply any C++ libraries.

This fixes the build on Mac OS X 10.6-10.8.

On Lion, force the use of macports-clang-3.4, which contains a fix for
a Lion-specific libc++ bug. The fix appears not to be present in
macports-clang-3.7 which is the compiler that would otherwise be used.

On Leopard and earlier, use the cxx11-1.1 portgroup's default behavior
of using gcc6 and its libstdc++.

Closes: https://trac.macports.org/ticket/55382