basiliskii: new version
this commit reinstates the basiliskii port to the
latest version that support the JIT compiler.
Also adds a separate basiliskii-gui port that
can be compiled with a separate architecture.

This allows basiliskii to be 32bit, and the GUI
to be the native architecture (usually 64 bit), and
therefore prevents rebuilding all of gtk2 as +universal

closes: https://trac.macports.org/ticket/18691