Update Portfile
This is an updated Portfile that should address some of the patches in the previous one.  A complication is that it uses meson build, which is a new autotools replacement.  BAli-Phy builds with clang++ on recent OSX versions now.  There is a working homebrew Formula at bredelings/homebrew-bioinformatics.  However, this goes beyond that because it uses the new build system.  This is the first Portfile I've tried to write, so any comments would be helpful.