lirc: depend on man's man2html instead
The standalone 'man2html' from the man2html port does not work for lirc.
The generated html versions of the manpages need to be generated with
the 'man2html' that comes with the man port. There are other pieces
of software out there calling themselves 'man2html', but lirc's
build system requires a particular implementation,
having e.g. the -M and -r options.