p5-dist-zilla-plugin-contributorsfromgit: add optional attribute 'target_path'
Allows limiting the list of contributors to committers to a specific
path in a git repository as opposed to the repository as a whole.  Useful
for listing contributors to individual members of macports/macports-port,
macports/macports-contrib and the like.

Default behavior, without setting this attribute, is to list all contributors
to the current repository as before.