glib2: Default to +quartz when -x11 is given
When building pango -x11 in a clean prefix, glib2 will fail to install,
because the -x11 variant is being passed down to glib2, but glib2
requires either +quartz or +x11, and +quartz was not enabled by default
when x11 was disabled.

This caused problems when implementing variant support for the buildbot
in https://github.com/macports/mpbb/pull/5 and initially caused me to
think that I would have to revert
  https://github.com/macports/mpbb/commit/f6e46815a5033358212ce383956fd567eb51e53c
which was added due to https://github.com/macports/mpbb/pull/4 and
https://lists.macports.org/pipermail/macports-dev/2017-June/035978.html.

This solution should instead work without the revert and still allow the
buildbot to build both wine and +quartz-x11 ports.

See: https://trac.macports.org/ticket/52742