app PortGroup : change executable to launch script
add default MacPorts expected PATH to launch script

by default the app PortGroup places a symlink to
the executable in the MacOS bundle. However, this
symlink fails to launch in many cases, because
the -psn parameter passed by default by MacOS is
rejected by the executable. Change the default
behaviour to making a simple bash launch script
instead, that does not pass -psn.

also - the executable expects the standard MacPorts
path setup to find associated applications it may need
to run. By default, replicate this PATH for the
MacOS bundled application as well