textproc/source-highlight: use cxx-1.1 PortGroup, force stdlib selection, revbump.
This selects the right compiler and stdlib, which is especially
important on older systems that need to use our stdlib based on a newer
GCC version that provides proper C++11 support.

Since libtool is picky about compiler flags and likes to strip unknown
flags out, force the stdlib selection to have libsource-highlight link
against the correct stdlib (incidentally also the one that boost was
built against).

Fixes: https://trac.macports.org/ticket/56013