pan2: update run time theme dependencies, rebuild
Unlike gtk3, the default theme for gtk2, Raleigh, is built in.

Although gnome-themes-standard provides additional gtk2 themes
Adwaita, Adwaita-dark and HighContrast, installing it is not required
for the default and no change in theme is provided without associated
configuration changes.

For the gtk3 build, depending on adwaita-icon-theme, which is a dependency
of gnome-themes-standard, is sufficient for the proper default theme display.