dia: remove unnecessary theme dependency
The default theme for gtk2, Raleigh, is built in.

Although gnome-themes-standard provides additional gtk2 themes
Adwaita, Adwaita-dark and HighContrast, installing it is not required
for the default and no change in theme is provided without associated
configuration changes.

No rebuild as there is no change in the build or running of dia.