wine*: Blacklist macports-clang-5.0 on OS X 10.6
We could also install cctools to get a more recent ranlib that would
understand this load command. However, this might as well be a bug in
clang to emit objects that the toolchain cannot understand and I am not
sure if the resulting binary could be loaded.

As I have no way to debug which load command this actually is,
let's try with an older version of clang.