virt-manager: Fix dependencies
Most libraries are not needed not for building as they are onlyaccessed
at runtime. Only list required ports to build virt-manager in
depends_build, move the rest to depends_run.

Remove incorrect gtk2 dependency, as it is in fact using gtk3, which is
already pulled in by spice-gtk and gtk-vnc. Update vte dependency
accordingly.