gcr: add variants for x11 and quartz
until now, gcr required gtk2 or gtk3 to be
installed +x11 (for UI code). This prevented
any ports requiring gcr from being installed
with gtk2 or 3 +quartz active.

This commit adds a +quartz variant to gcr
that disables the UI, and therefore no need for
the +x11 code. This allows ports that use gcr
to be installed +quartz.

A default +x11 variant is also added to mimic
existing behaviour.

closes: https://trac.macports.org/ticket/41839