ntpsec: Fixes build when default Python is old.
A Python program used as a build helper was being run with the
system default Python, and failing when that was too old.  This
fix (possibly acceptable upstream) causes it to use the same
Python as is being used to run waf, which is already forced to
be the MacPorts Python 2.7.

TESTED:
On both MacPro/10.9 and PowerBook/10.5, with python_select set to
python25-apple, verified that the build failed without this fix
and succeeded with it.  Also tested with the usual python27.