openscad: fix Resource directory discovery
openscad finds the resource directory by searching a preset
path list looking for a folder "color-schemes" so this needs
to be in the resource directory.

Fixes missing examples from splash window and allows other items
in the resource directory to be discovered.

Finally, move the fonts and locales folder into the resource directory
as well so that all will be in the same location.