glibmm: Fix character set conversion with libc++
glibmm contains code that attempts to convert between std::wstring and
glibmm's Glib::ustring string class, which stores characters in UTF-8.
This conversion code was using a fallback solution that eventually calls
iconv to convert from the "WCHAR_T" character set to UTF-8.

On macOS, iconv --list does not show WCHAR_T as supported character set
and comments in the libiconv code handling WCHAR_T suggest it was
written for glibc, which is not used on macOS. Consequently, the
conversion failed, which caused Glib::ustring to throw an exception.

This exception was not caught by inkscape and caused a crash when
starting inkscape, rendering it useless.

Fortunately, glibmm already contained code to deal with libc++'s
std::wstrings, but it was guaded by a check for the __STDC_ISO_10646__
prepocessor define. Allowing _LIBCPP_VERSION in addition to
__STDC_ISO_10646__ makes code compiled against libc++ use this code path
and solves the issue.

Closes: https://trac.macports.org/ticket/56214