docbook-xsl: Use the non-namespaced version
Fixes rendering of manpages by xsltproc.

Closes: https://trac.macports.org/ticket/56354

When the port was updated to 1.79.2 in 04e6959f23bb6b1bd40f40f4d1a3ee2fb43e0714,
it was inadvertently updated to the namespaced version, because upstream
changed the default packaging from non-namespaced to namespaced.

The namespaced version can be found in the separate docbook-xsl-ns port,
which still needs to be updated to the latest version, and ideally
merged into the docbook-xsl port as a subport. Renaming the ports and
updating the dependents might also still be considered.