opensc: remove +doc variant, add stylesheet dir
Since all build-time dependencies that once were conditional to building
the documentation are now required to build the man pages, remove the
+doc variant and always build the docs. This also has the benefit of
simplifying the possible configurations and reducing the test matrix.

In addition, the stylesheet directory is now explicitly set since the
install location has changed and is no longer auto-detected by the
configure script.

Fixes: https://trac.macports.org/ticket/56512