octave: Complete update to version 4.4.0
This update includes four hacks:

* Quitting from the GUI now works.  This is achieved by adding a
`printf(3)` in the quit handler. It would appear that there's a
buffering issue somewhere, perhaps a missing `fflush(3)`. This
should probably be looked into further or at least reported upstream.
It seems unlikely to be specific to macOS and could cause hangs in
other contexts besides quitting.

* The command `__opengl_info__` (five underscores) should now work
the first time it's run. There's a known race condition which is
worked around by delaying querying OpenGL info.

* GNU `tar` is added as a run-time dependency, rather than fixing
the test suite by BSD `tar`.

* FLTK variant is commented out.  As noted in a previous comment,
FLTK support in Octave appears to be on its way out.

In addition to the above hacks, Octave's PATH is prefixed with
${prefix}/bin.  This fixes an issue where running `help` in the
Octave GUI would generate errors.

With these updates, the GUI works as expected and the test suite
runs with no unexpected errors: Summary:

  PASS                            14912
  FAIL                                0
  XFAIL (reported bug)               54
  SKIP (missing feature)             76
  SKIP (run-time condition)          28

One final note:  It appears that the HDF5 data library isn't found
by Octave configure if  hdf5 is compiled with variant `openmpi'.
So if openmpi is a default variant, try disabling it when building
hdf5 and before building octave.