wireshark2: Force reindex by applying some change
The wireshark24 port identified as wireshark2 before and was added to
PortIndex with that name. Although the name of wireshark24 was fixed in
879e68cd17, the wrong entry for wireshark2 stayed in PortIndex. The
wireshark2 Portfile also needs to be changed to force a reindex to
replace the wrong entry in PortIndex.

Removing revision as 0 is the default.

See: https://trac.macports.org/ticket/56459
See: https://lists.macports.org/pipermail/macports-dev/2018-May/038915.html