compilers portgroup: update compilers.fortran_variants
compilers.setup changes compilers.variants based on which compilers have
been chosen, but it neglected to also update compilers.fortran_variants
in the same way. This led to incorrect information being returned by at
least the fortran_variant_name proc, in the case that a variant was
requested but didn't actually exist in the port.

Fixes: https://trac.macports.org/ticket/56582