vim, MacVim: Do not link with -framework Python
Using -framework for linking is not correct, as MacPorts provides
multiple versions of Python. This would always pick up the 'port select'
version and not the one request by variants. Restore parts of the patch
for python that were errorneously removed in 00f3dc13b8 and 0dbff7850a.

Incrementing the revision as this change means that the ports might have
linked against the wrong version of python, for example against the
macOS system framework instead of a version provided by MacPorts.

Closes: https://trac.macports.org/ticket/56652