gnome/gstreamer1-gst-plugins-good: force-enable features if selected via variants.
Just removing the --disable-foo flag will make configure auto-detect
feature foo, which is not what we want, if the variant is enabled.
Instead, it should be force-enabled and configure fail, if the feature
is not usable (for instance because we forgot to add proper dependencies
and the port is built in trace mode or the build system just lacks the
required dependencies because it's mostly clean).