Remove maintainer gmail.com:michel.sylvan
The only reason he was added as maintainer of this port in 2011 in
https://trac.macports.org/ticket/31501 was because he had originally
submitted Python 2.4 and 2.5 ports for this module in 2008 in
https://trac.macports.org/ticket/14838.

He is in fact a MacPorts committer with the handle hircus; his last
(only?) commit to MacPorts was a66aebd267f48b8a9de61d07dcb05852d40a7c7b
in late 2008.

He was moved to the Retired section of the MacPortsDevelopers wiki page
in 2017.