ntpsec: update to 1.1.1
This includes the changes for compatibility with macOS<10.13,
which can also be seen at:

      https://gitlab.com/fhgwright/ntpsec/tree/macports_1_1_1

Defers shebang rewrite to destroot copy, for better consistency
between patched sources and quasi-upstream version.  Also, if
upstream ever provided shebang rewriting, it almost certainly
would not apply to the original sources.

NOTE:
The "upstream fork" containing the fixes for MacPorts has now
been reorganized to use a single branch with tags, rather than
a per-release branch.  The branch can be found at:

	https://gitlab.com/fhgwright/ntpsec/tree/macports-releases

In particular, note that the 1.1.0 version formerly found at
the NTPsec_1_1_0_macports branch (and mentioned in the commit
message for 31b309b046) is now found on the macports-releases
branch at the macports_1_1_0 tag.

TESTED:
Built and ran on MacPro 10.9, MacBookPro 10.9, PowerBook 10.5,
and VMs for 10.5-10.13.
Also verified that build works with default Python set to 2.4,
confirming that the "old Python" fix still works.

Closes: https://trac.macports.org/ticket/56657