Update to sbcl-1.4.9
The SBCL bootstrap binary had to be updated as well, so it is
suspected that the PowerPC version of this port will no longer build.
Advice about either possible deprecation or updates in this manner are
solicited.

The `lang/maxima` port has removed their `revision` keyword, so this
commit does not include an update to `lang/maxima`.  A query was sent
to recent Maxima port commiter for clarification.