netperf: enable demo mode
Also fix the port to use github portgroup properly. The patch file is a
partial patch from 40c8a0f of the upstream repository.

Demo mode is required for use "flent", a tool to visualize the output of
netperf.

Note that configure argument to this program affect its output, formats,
and other things that really should be run-time toggles. Be careful when
selecting them.