p5-config-simple: use default livecheck
Livecheck doesn't work for this module due to indexing
issues with the current version making it invisible to livecheck.

This leads to the livecheck error message

Error: livecheck failed for p5-config-simple: extracted version '4.58' is older than livecheck.version '4.59'

when, in fact, version 4.59 exists and fetches without problems.

This is a long standing problem with no response from upstream to fix the distribution so that it indexes
properly on metacpan.org. When and if this happens default livecheck should work properly. Note version 4.59
contains an important bug fix so reverting to verion 4.58 is not recommended.

See https://rt.cpan.org/Public/Bug/Display.html?id=113751